pyproject.toml
[tool.poetry]
name = "ot-harjoitusty"
version = "0.1.0"
description = ""
authors = ["Iltsukka <113025832+Iltsukka@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^2.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
coverage = "^7.4.4"
pylint = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
